requirement.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
langchain-experimental==0.3.0
transformers==4.44.2
torch==2.4.1
faiss-cpu==1.8.0
sentence-transformers==3.1.0
openai==1.45.0
groq==0.11.0
python-dotenv==1.0.1
huggingface-hub==0.25.0
